Run ark client container as the active user
Since the ark container runs as a non-privileged user by default there are cases where it may not be able to read some user's config files if the permissions are more restrictive. Running the ark as the active user will make sure that the config file can be used in all cases.

Signed-off-by: Carson Anderson <ca@carsonoid.net>