design doc template: remove status field
After discussing with the team, it doesn't make sense for all design proposals to have a draft/accepted status. Currently draft proposals can be kept open as PRs and all merged proposals are considered to be accepted. Removing the status field here removes the extra step of having to update the proposal status before merging.

Signed-off-by: Adnan Abdulhussein <aadnan@vmware.com>