Fix Image Handling
This change introduces a new target `build-env` that checks if there is
already a build-image in the docker-cache. If there is we use it. If not
we call build-image.

A user can still call make build-image and force a refresh of the image
if the build-image file is different.

This work is the first part of a set of changes that will improve the
way we manage the build-image.

Signed-off-by: Duffie Cooley <cooleyd@vmware.com>