Fix make clean target.
Make clean removes the .go directory from the current workdir.
The .go directory has the go mod cache when we are using containerized
builds. One doesn't simply rm a mod cache :) This change evaluated
whether we have a build image locally and if we do we use it to run
`go clean --modcache` to clean up the mod cache from .go.
Then we can delete the dir and the image.

Signed-off-by: Duffie Cooley <cooleyd@vmware.com>