Add namespace parameter in restic's code
This intends to be able to target specific namespaces when
having non-admin privileges.

When running velero as a non-admin user, listing pods from
other namespace is forbidden. The use of an environment variable in
the docker-compose.yml allows to target a specific namespace in which
we want to list pods.

Signed-off-by: Sebastien FELIX <sebastien.felix.ext@orange.com>