Add option to bypass calls to pvLister
This intends to bypass calls to pvLister in function
GetVolumeDirectory when we know all volumes are CSI.

If running velero as a non-admin user on a cluster, the call
to pvLister can cause errors for namespaces which are not
assigned to the user. If all volumes are CSI, there is no
need to use pvLister, hence the bypass.

Signed-off-by: Sebastien FELIX <sebastien.felix.ext@orange.com>